Move robot_constants to basic gopigo3 class
